WIP move net.imglib2.ui classes
TransformEventHandler -> bdv
